Changed dataset iter return type
Changed iteration over TabularDataset to return TabularRecord instances
