+++ prive/datasets/__init__.py
@@ -1,3 +1,2 @@
 from .dataset import Dataset
-from .dataset import TabularDataset
-from .record import TabularRecord
+from .dataset import TabularDataset, TabularRecord
